IDEcontext/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openclaw"
version = "0.1.0"
description = "Self-hosted AI agent — personal digital assistant with messaging UI and MCP server"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
keywords = ["ai-agent", "mcp", "self-hosted", "telegram", "slack", "whatsapp"]

dependencies = [
    "mcp>=1.9.0",
    "openai>=1.60.0",
    "python-dotenv>=1.0.1",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "click>=8.1.7",
    "rich>=13.9.0",
    "watchfiles>=1.0.0",
    "python-telegram-bot>=21.9",
    "slack-bolt>=1.21.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "httpx>=0.28.0",
    "python-frontmatter>=1.1.0",
    "pyyaml>=6.0.2",
    "croniter>=5.0.1",
    "anyio>=4.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[project.scripts]
openclaw = "openclaw.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.10"
strict = true
